# Conflicts:
#	Gateway/Request/BasicParameter/AdditionalParametersDataBuilder.php
#	Gateway/Request/BasicParameter/AmountDebitDataBuilder.php
#	Gateway/Request/BasicParameter/CurrencyDataBuilder.php
#	Gateway/Validator/AreaCodeValidator.php
#	Gateway/Validator/IssuerValidator.php

+++ Cron/LogCleaner.php
@@ -1,13 +1,12 @@
 <?php
-
 /**
  * NOTICE OF LICENSE
  *
  * This source file is subject to the MIT License
  * It is available through the world-wide-web at this URL:
  * https://tldrlegal.com/license/mit-license
- * If you are unable to obtain it through the world-wide-web, please send an email
- * to support@buckaroo.nl so we can send you a copy immediately.
+ * If you are unable to obtain it through the world-wide-web, please email
+ * to support@buckaroo.nl, so we can send you a copy immediately.
  *
  * DISCLAIMER
  *
@@ -33,14 +32,52 @@ use Magento\Framework\Exception\FileSystemException;
 
 class LogCleaner
 {
+    /**
+     * @var LogResourceModel
+     */
     private $resource;
+
+    /**
+     * @var Account
+     */
     private $accountConfig;
+
+    /**
+     * @var ResourceConnection
+     */
     private $resourceConnection;
+
+    /**
+     * @var Log
+     */
     private $logging;
+
+    /**
+     * @var DirectoryList
+     */
     private $directoryList;
+
+    /**
+     * @var File
+     */
     private $driverFile;
+
+    /**
+     * @var IoFile
+     */
     private $ioFile;
 
+    /**
+     * Log Cleaner constructor
+     *
+     * @param LogResourceModel $resource
+     * @param Account $accountConfig
+     * @param ResourceConnection $resourceConnection
+     * @param Log $logging
+     * @param DirectoryList $directoryList
+     * @param File $driverFile
+     * @param IoFile $ioFile
+     */
     public function __construct(
         LogResourceModel $resource,
         Account $accountConfig,
@@ -59,6 +96,9 @@ class LogCleaner
         $this->ioFile              = $ioFile;
     }
 
+    /**
+     * Cron that clean the logs after specific period
+     */
     public function execute()
     {
         $retentionPeriod = (int) $this->accountConfig->getLogRetention();
@@ -75,6 +115,12 @@ class LogCleaner
         return $this;
     }
 
+    /**
+     * Delete logs from data base
+     *
+     * @param int $retentionPeriod
+     * @return void
+     */
     private function proceedDb(int $retentionPeriod)
     {
         try {
@@ -87,6 +133,13 @@ class LogCleaner
         }
     }
 
+    /**
+     * Delete files that contains logs
+     *
+     * @param int $retentionPeriod
+     * @return void
+     * @throws FileSystemException
+     */
     private function proceedFiles(int $retentionPeriod)
     {
         if ($files = $this->getAllFiles(DIRECTORY_SEPARATOR . 'log' . DIRECTORY_SEPARATOR . 'Buckaroo')) {
@@ -94,16 +147,20 @@ class LogCleaner
             foreach ($files as $file) {
                 $fileInfo = $this->ioFile->getPathInfo($file);
                 $fileName = $fileInfo['filename'];
-                $matches = null;
-                if (preg_match('/[\d]{4}\-\d{2}\-\d{2}/', $fileName, $matches)) {
-                    if (strtotime($fileName) <= $retentionTime) {
-                        $this->driverFile->deleteFile($file);
-                    }
+                if (preg_match('/[\d]{4}\-\d{2}\-\d{2}/', $fileName)
+                    && (strtotime($fileName) <= $retentionTime)) {
+                    $this->driverFile->deleteFile($file);
                 }
             }
         }
     }
 
+    /**
+     * Get all files from log directory
+     *
+     * @param string $path
+     * @return array
+     */
     private function getAllFiles(string $path): array
     {
         $paths = [];
